move utilities in cfchecks, datachecks and missing files. update tests and docs
