+++ docs/checks.rst
@@ -9,13 +9,13 @@ Missing values
 ==============
 The following functions can be used to mask missing values in :class:`Indicator` outputs.
 
-.. autofunction:: xclim.core.checks.missing_any
+.. autofunction:: xclim.core.missing.missing_any
    :noindex:
 
-.. autofunction:: xclim.core.checks.missing_pct
+.. autofunction:: xclim.core.missing.missing_pct
    :noindex:
-.. autofunction:: xclim.core.checks.missing_wmo
+.. autofunction:: xclim.core.missing.missing_wmo
    :noindex:
 
-.. autofunction:: xclim.core.checks.at_least_n_valid
+.. autofunction:: xclim.core.missing.at_least_n_valid
    :noindex:
